finish embark
